+++ reversion/signals.py
@@ -1,5 +1,13 @@
 from django.dispatch.dispatcher import Signal
 
 
-pre_revision_commit = Signal(providing_args=["instances", "revision", "versions"])
-post_revision_commit = Signal(providing_args=["instances", "revision", "versions"])
+_signal_args = [
+    "instances",
+    "ignore_duplicates",
+    "meta",
+    "revision",
+    "versions",
+]
+
+pre_revision_commit = Signal(providing_args=_signal_args)
+post_revision_commit = Signal(providing_args=_signal_args)